src/requirements.txt
python-binance
python-dotenv
python-telegram-bot
backtrader
ccxt
git+https://github.com/wilfredbtan/bt-ccxt-store.git
pandas
termcolor
coloredlogs
